Postpone inclusion of editor-annotation.ttl
From discussion on the UCO Ontology Committee call on 2020-04-27, a
sufficient group of those present agreed that the documentation matter
of UCO Issue 178 would need to be more extensively addressed, in a
timeline that would significantly and negatively impact the release of
version 0.4.0.

This patch removes editor-annotation.ttl and its usage.  The patch is a
result of, but does not completely address, UCO Issue 178:
https://github.com/ucoProject/UCO/issues/178

Usage was only found in vocabulary.ttl, from running `grep -Ri` for two
string patterns:

* `editor-annotation`
* `edt:`

This patch was tested for normalized content with the same methodology
as in Pull Request 191:
https://github.com/ucoProject/UCO/pull/191

There is no overlap in effect between that Pull Request and this patch.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>