Normalize content with rdf-toolkit
This patch applies rdf-toolkit to all of the Turtle files in UCO
(including *-da.ttl files), following the normalization process adopted
by the CASE community.

Normalization was confirmed by the CASE-Examples-QC repository.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of the instructions to this
toolkit is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software or toolkit is necessarily the best
available for the purpose.

References:
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC
  Branch: OC-43
* [CASE ONT-5] (CP-3) Establish ontology normalization procedure
* [Normalization process] https://github.com/casework/CASE/blob/master/NORMALIZING.md

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>