Normalize syntax
This applies rdf-toolkit following the nomralization directions of the
CASE CONTRIBUTE.md file.  This is in addition to the prior patch that
normalized observable.ttl, which was suggested as needing attention.
The CASE-Examples-QC repository later identified other files needing
normalization.

No manual content changes were made in this patch.

References:
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC
* [OC-63] UCO Version 0.6.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>