Consolidate versionInfo to one declaration of 0.6.0
This patch bumps the version declaration of UCO to 0.6.0.  It also
removes some unmaintained versionInfo declarations from other
ontologies.  There is a plan to restore these other versionInfo
declarations in CASE CP-17, but CASE and UCO are currently analyzing
options for the version-incrementing strategy.

This patch is confirmed to not cause normalization issues.

References:
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC
* [CASE ONT-64] (CP-17) Define scheme for CASE and UCO importable
  versioned IRIs

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>