Base UCO imports on versionIRI property
This patch takes the versionIRI practices the CASE Ontology Committee
voted on and accepted in their CP-26, and applies them to UCO 0.4.0 in
support of UCO CP-40.  UCO CP-40 is a parallel proposal to CASE CP-26.

A serving domain of https://archive.unifiedcyberontology.org is assumed,
but has not been discussed in the UCO Ontology Committee and may be
subject to change.

rdf-toolkit.jar was NOT used to normalize the files, because UCO 0.4.0
did not apply a content normalization step before release (like was
later done in UCO 0.5.0 and 0.6.0).  However, rdf-toolkit.jar was run
against each turtle file to confirm no syntax or missing-prefix issues
were introduced in this patch.

References:
* [CASE ONT-412] (CP-26) Implement versionIRI in CASE 0.2.0
* [OC-101] (CP-40) Implement versionIRI in UCO 0.4.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>