Pass ontological subclass relationships through extra pyshacl argument
`rdfs:subClassOf` statements are not recognized when included in the
graph passed to `pyshacl`'s flag `--shacl`.  They *are* recognized when
passed via the `--ont-graph` flag; see pyshacl issue 14.  Hence, to get
subclass awareness (as would apply to, e.g., all Facet subclasses in
UCO), the monolithic UCO file must be passed to `pyshacl` in two of its
flags.

This patch is still expected to fail CI, as data corrections are coming
in a follow-on patch.

References:
* [OC-68] (CP-23) Convert current property restrictions and domain
  assertions to SHACL class shapes
* [OC-159] pyshacl sh:class constraints seem to not recognize
  subclassing
* [pySHACL 14] https://github.com/RDFLib/pySHACL/issues/14