Normalize ontology files after adding type constraints
This causes a significant degree of line churn due to the PropertyShape
nodes getting lexicographically-lower properties.

With this patch, CI based on known induced errors in example data
passes.  A further manual fix is to be applied, and another CI component
based on property shapes related by subclasses is to be added.

References:
* [OC-68] (CP-23) Convert current property restrictions and domain
  assertions to SHACL class shapes

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>