Add test file representing a result-node of an Action
Testing a CASE-style example of an action that resulted in a file, a
curious error arose claiming the file was not a UcoObject.

This UCO test distills the CASE example down to UCO concepts, and
attempts an exhaustive listing of subclass combinations for the more
general ObservableObject.  CI will report a validation error, which I've
recorded as the "PASS" validation graph.  When the behavior is
corrected, the graph can be updated.

References:
* [OC-68] (CP-23) Convert current property restrictions and domain
  assertions to SHACL class shapes