Add integers to XFAIL for any solution to semi-open vocabularies
The Make-managed SHACL validation results are also regenerated in this
patch.

v1 of the implementation suggested some use of `sh:not`.  Attempting to
use this to allow a named vocabulary or another named vocabulary is
accidentally too broad an acceptance range.  Integers can be accepted
if, instead of `xsd:string`, we use `sh:not [ vocabulary:XVocab ]`.

This patch adds integers to the XFAIL test to ensure we don't make this
error in any revisions to the strategy.

References:
* [OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree with
  its implementation with owl:oneOf

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>