Allow sh:Info-level violations to still be considered conformant
UCO CP-100 is the first proposal to make use of shapes of a severity
other than the default `sh:Violation`.  This patch allows an affirmative
SHACL conformance result with all non-Violation-level validation
results.

The flag used on `pyshacl` is `--allow-warnings`.  At the time of this
patch, `--allow-infos` is implemented, but not yet released (see pySHACL
PR 127).

The tests for semi-open vocabularies are accordingly revised to move the
Info-level validation results into the PASS sample file.

The Make-managed validation results files are regenerated in this patch.

References:
* [OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree with
  its implementation with owl:oneOf
* https://github.com/RDFLib/pySHACL/pull/127

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>