Remove vocabulary references, and datatype constraint, from core:Relationship
The strategy deployed on `core:kindOfRelationship` at the
`core:Relationship` level had an unrealized effect: No vocabulary
defined in the style of UCO's relationship vocabularies, but outside the
scope of UCO, would be usable.

It is unclear how to resolve this issue fully.

This patch attempts some preservation of vocabulary constraint at the
appropriate semantic level, at the exchange of being unable to make any
`sh:datatype` constraints for `core:kindOfRelationship` at the level of
`core:Relationship`.

The unit test of what was previously the full feature has been split
into a PASS-ing partial and XFAIL-ing full test.

A follow-on patch will regenerate Make-managed files.

References:
* [OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree with
  its implementation with owl:oneOf

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>