Remove vocabulary references, and datatype constraint, from core:Relationship
The strategy deployed on `core:kindOfRelationship` at the
`core:Relationship` level had an unrealized effect: No vocabulary
defined in the style of UCO's relationship vocabularies, but outside the
scope of UCO, would be usable.

It is unclear how to resolve this issue fully.

This patch removes all attempts at constraining vocabularies for
`core:kindOfRelationship`, awaiting a re-design of the implementation of
semi-open vocabularies.

The unit test of what was previously the full feature has been split
into a PASS-ing partial and XFAIL-ing full test, for both the
`relationship_PASS` and `_XFAIL` samples.

A follow-on patch will regenerate Make-managed files.

References:
* [OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree with
  its implementation with owl:oneOf

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>