Remove vocabulary references from core:kindOfRelationship
This patch finishes reverting the addition of vocabulary references to
`core:kindOfRelationship` within the `core` namespace.  As a
side-effect, the vocabulary import added as part of the OC-12/CP-100
patch series is no longer needed, and thus is also reverted.

This change should have been applied as part of 3ab9126, but was missed.

No effect was observed on Make-managed files.

References:
* [OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree with
  its implementation with owl:oneOf

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>