Add OWL-SKOS mix-in class
This patch adds initial support for SKOS taxonomies as a
hierarchy-enabled alternative to UCO's current string-vocabularies
practice.

While the typical SKOS namespace is prefixed, the import of SKOS is
done by referencing an OWL-DL compatible subset of SKOS as a version IRI
import.  (See SKOS Reference C.3.)

Paul Brandt identified the SKOS strategy used in this patch, though his
initial draft was in another repository.  I ported the property to UCO's
ontology repository due to needing to satisfy references for SHACL.

This patch is a reduction from the first definition of
`core:TaxonomicConcept` drafted for UCO CP-99.

References:
* [OC-140] (CP-99) UCO should provide a SKOS taxonomy of device types
  for observable:deviceType
* https://github.com/ucoProject/UCO/issues/363
* https://www.w3.org/TR/skos-reference/#namespace-documents
  Section C.3, "SKOS RDF Schema - OWL 1 DL Sub-set (informative)"

Co-authored-by: Paul Brandt <paul@brandt.name>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>