Extend Dublin Core Terms FileFormat class to be OWL Class
Dublin Core's `dcterms:FileFormat` class is an existing class that
represents file formats, including but not limited to IANA Media Types.
(The reference to IANA Media Types is found on the `dcterms:format`
property, where the IANA list is recommended, but not required.)

`dcterms:FileFormat` is defined as an `rdfs:Class`.  In order for UCO,
as an OWL ontology, to use this class as a range of a property
(especially `observable:mimeType`), it needs to be designated as an
`owl:Class`.

Note there are some things this patch does not do:

* This patch does not import Dublin Core.  It appears Dublin Core's
  model is incompatible with OWL 2 DL.  For instance, the property
  `dcterms:format` appears to violate the OWL 2 DL separation of
  datatype properties (range of literals) from object properties (range
  of objects), on review of its sub-property `dcterms:extent` having a
  range of literals.
* This patch does not extend `FileFormat`'s superclasses to also be
  OWL Classes.
* This patch does not adapt the `dcterms:format` property noted above,
  because in addition to its OWL 2 DL compatibility issues, its semantic
  scope extends beyond data formats.

Compatibility notes:

Extending `dcterms:FileFormat` to be both an OWL Class and RDFS Class is
compatible with OWL 2 DL.  Reviewing this record:
https://www.w3.org/TR/2012/REC-owl2-mapping-to-rdf-20121211/#Parsing_of_the_Ontology_Header_and_Declarations
Table 5
Row 2

An ontology that also designates `dcterms:FileFormat` as an `rdfs:Class`
would have the pair of `owl:Class` and `rdfs:Class` statements reduced
to `owl:Class` only as part of its OWL parse.

If an ontology has chosen to import Dublin Core as a whole, it has
already implicitly made a commitment to use OWL 2 FULL instead of
OWL 2 DL.

References:
* https://github.com/ucoProject/UCO/issues/363
* https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#FileFormat
* https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#format

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>