Revert RDF list reference-consolidation and test for alignment
This patch undoes an engineering convenience put in place as part of UCO
CP-100.  RDF Lists that were part of semi-open vocabularies were given
IRIs, so they could be referenced for OWL datatype definitions and for
SHACL membership testing.  This was acknowledged as an incompatibility
with OWL 2 DL, which requires that RDF Lists be identified as blank
nodes.  The concepts were intended to remain until an OWL test mechanism
would identify this error.

A test mechanism is now under development as part of UCO Issue 406, and
correctly flags IRI-identified RDF lists.  Hence, this patch undoes the
change.

To ensure the RDF lists are kept in sync. across their duplicate
locations, a Python unit test has been added to confirm list-equality.

References:
* [UCO OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree
  with its implementation with owl:oneOf
* https://github.com/ucoProject/UCO/issues/406

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>