Cut errant rdf:list artifact
Somewhere in the chain between core rdflib, pyshacl, and
rdf-toolkit.jar, the normalized Turtle content sometimes picks up
redundant list artifacts that vary between runs.  The anonymous list cut
in this patch has been seen to waver in how much of it is left as a
detached blank node.

This patch removes the orphaned list.  Guidance needs to be provided
that, as long as this bug in the tool stack persists, the Make-managed
generated output should not have it Git-tracked.

References:
* https://github.com/ucoProject/UCO/issues/406

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>