Remove usage of sh:declare
SHACL Specification Section 5.2.1 specifies a "viral" behavior of
`sh:declare` throughout an OWL transitive closure.  This patch removes
usage of `sh:declare` as a matter of lack of authority for non-UCO
prefixes.  It just so happens the only place this was used was in the
introduction of the OWL SHACL review mechanisms of Issue 406.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ucoProject/UCO/issues/406
* https://github.com/ucoProject/UCO/issues/457

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>