Normalize styles
This patch is a manual adjustment to apply UCO current ontology spelling
and linking practices.

* While the `rdfs:label` for the `owl:Ontology` includes the `uco-`
  prefix on the namespace short form, the prefix encoded in Turtle for
  concepts has not included the `uco-` prefix.
* `rdfs:seeAlso` has been used when drafting ontology concepts in files
  named `drafting.ttl`, as a part of CASE development practice, to tie
  the concept to its originating work-anchor (whether Jira ticket,
  Github Issue, or other).  However, by the time concepts migrate into
  the UCO or CASE ontology Turtle files, `rdfs:seeAlso` is not used to
  reference originating tickets.

The first issue caused a breakage in normalization, because the
`@prefix` declaration didn't quite agree with all of the UCO concept
references.  So, this patch applies the manual spelling changes, and a
follow-on patch will normalize the new `configuration.ttl` file.

References:
* https://github.com/ucoProject/UCO/issues/432

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>