Restore isConfigurationOf; add usesConfiguration
Sean removed `tool:ConfiguredTool` and `tool:isConfigurationOf` from the
implementation for Issue 405, without explanation.  His revised
demonstration of how to represent a tool being used with a configuration
within an action included this excerpt (not complete UCO JSON-LD, but
sufficient to demonstrate the design issue):

```json
[
    {
      "@id": "kb:relationship-d6d18770-cb60-4779-b28a-905846dbd3f4",
      "@type": "uco-core:Relationship",
      "uco-core:source": "kb:AnalyticTool-DAE5EE58-E5ED-4588-93BE-CDEC6FAA9C6A",
      "uco-core:target": "kb:Configuration-4bc42c81-e9e7-4f48-b066-17e7fdca0f92",
      "uco-core:kindOfRelationship": "Has_Configuration",
      "uco-core:startTime": "2010-01-15T17:59:43.25Z",
    },
    {
      "@id": "kb:Analysis-7cd51fa7-63ee-4f40-a482-9ce8333c7556",
      "@type": "uco-analysis:Analysis",
      "uco-core:name": "compute string similarity",
      "uco-action:instrument": [
        "kb:AnalyticTool-DAE5EE58-E5ED-4588-93BE-CDEC6FAA9C6A",
        "kb:relationship-d6d18770-cb60-4779-b28a-905846dbd3f4",
        "kb:Configuration-4bc42c81-e9e7-4f48-b066-17e7fdca0f92"
      ]
    }
]
```

`tool:ToolConfigurationTypeFacet` was also removed, and I agree that
that is obviated with the `Configuration` class.  Because that `Facet`
is removed, another mechanism is needed to link a configuration to a
`Tool` or a `Software`.

This patch restores `ConfiguredTool`, and adds a similar class
`ConfiguredSoftware` to meet the expanded objective of Issue 432.

`isConfigurationOf` is restored as a more generically-specified
property, leaving type-scoping to the SHACL associations.

`configuration:usesConfiguration` is added to link a `ConfiguredX`
instance to a `Configuration`.

With the above new and restored properties and classes, the above
JSON-LD snippet is now expressible like this, most notably removing the
complexity of references in `action:instrument`:

```json
[
    {
      "@id": "kb:configured-tool-c6718138-e953-4731-bf52-d7d89eae9c85",
      "@type": [
          "uco-tool:AnalyticTool",
          "uco-tool:ConfiguredTool",
      ],
      "uco-configuration:isConfigurationOf": "kb:AnalyticTool-DAE5EE58-E5ED-4588-93BE-CDEC6FAA9C6A",
      "uco-configuration:usesConfiguration": "kb:Configuration-4bc42c81-e9e7-4f48-b066-17e7fdca0f92",
      "uco-core:objectCreatedTime": "2010-01-15T17:59:43.25Z",
    },
    {
      "@id": "kb:Analysis-7cd51fa7-63ee-4f40-a482-9ce8333c7556",
      "@type": "uco-analysis:Analysis",
      "uco-core:name": "compute string similarity",
      "uco-action:instrument": "kb:configured-tool-c6718138-e953-4731-bf52-d7d89eae9c85"
    }
]
```

Tests are updated to exercise the new classes and properties.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ucoProject/UCO/issues/405
* https://github.com/ucoProject/UCO/issues/432

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>