Require all object references in UCO be to non-blank nodes
This patch is a find-and-replace to reset the value of `sh:nodeKind` in
SHACL constraints, only done within the scope of the UCO ontology files.
The non-UCO shapes files (i.e. for Collections Ontology, OWL) are
intentionally left unaffected by this change, to allow for
interoperability with non-UCO content using those ontologies' concepts.

A follow-on patch will regenerate Make-managed files.

This patch will fail CI from the adjusted constraint.  A further
follow-on patch will correct the unit tests (separated as demonstration
of impact).

References:
* https://github.com/ucoProject/UCO/issues/430

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>