Add tests to prepare for review of linking to prior versions
The shape `uco-owl:versionIRI-multiversion-shape` also makes a reference
to "the import closure of [ontology] O".  This patch assumes that it is
more valuable to report multiple values of `owl:versionIRI` on any
`owl:Ontology`, regardless of whether those versions are both imported
by some in-common ontology.

A follow-on patch will generate Make-managed files.

References:
* https://github.com/ucoProject/UCO/issues/437

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>