Bump versionInfo; add priorVersion compatibilities; adjust imports in tool.ttl
`tool.ttl` uses a concept from the identity namespace without importing
it.  This patch removes the core import, as redundant with the
configuration import, and adds an import for the identity namespace.

Note that due to most namespaces importing core, and core being
backwards-incompatible with 0.9.1's core, all UCO ontologies except for
`vocabulary.ttl` are backwards-incompatible with their respective 0.9.1
ontologies.  `vocabulary.ttl` is incompatible for its own reason.

References:
* https://github.com/ucoProject/UCO/issues/437

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>