Integrate catalog construction into CI chain
The construction script now handles multiple input ontology files, but
with the requirement that they be in the same directory.

Interfaces have also been added to handle imported, possibly non-CDO
ontology references in two ways:
* With a TSV file mapping ontology IRIs or version IRIs to files.
* With optional references to (effectively imported) `catalog-v001.xml`
  files.

Another behavior change is implemented: the focus ontologies are now
also added to the `catalog-v001.xml` file, in part to support when
multiple graph files are in one directory, and in part to support
re-consumption of `catalog-v001.xml` by the `catalog-v001.xml`
generating script.

The rationales for how to handle ontology--file mappings outside the
scope of UCO (in both upstream and downstream directions) include:
* Symbolic links could have been used to pool all file references into
  the `/dependencies` directory.  Windows users that run `git clone`
  without symbolic links enabled for their system would encounter
  significantly counter-intuitive errors.
  - This also would not iterate well with consumers of the catalog
    script outside of UCO (e.g. CASE).
* A Makefile could have been made to normalize the dependent ontology
  files into the same Turtle style (or even away from RDF-XML, which the
  Collections Ontology currently uses as sole format).  However, this
  would again be a point of difficulty for Windows users, as they would
  have to run `make` to create the files referenced in the catalog XML.
* Copying files into a Git repository introduces code-drift issues that
  are difficult to manage.  When the copied files were themselves
  tracked in Git, this is counter to the purpose of Git submodules.

This patch goes on the assumption that Git submodules and recursive
cloning are a reasonable minimal requirement to access full local-file
ontology interaction.

The catalog generating script in this patch state has been tested
(offline) with CASE and CASE-Corpora as users, via a submodule chain
starting with CASE-Corpora.

The `CONTRIBUTE.md` file has also been updated to add usage
documentation, and to fix a copy-paste error from some time ago.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ucoProject/UCO/issues/449

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>