Revert direct Pattern namespace import
The previous patch included a direct Pattern namespace import.  This
turns out to not be necessary because the Pattern namespace is already
in the transitive `owl:imports` closure, via the Action namespace.
Hence, the `Pattern` class is already available for `ObservablePattern`.

I realized this when I noticed the `catalog-v001.xml` file didn't
require an update from adding the direct import.

References:
* https://github.com/ucoProject/UCO/issues/543

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>