Add test that sh:datatype does not reference an OWL Class
This test is spelled as a "Negative" constraint to avoid attempting to
hard-code a list of the recognized OWL "built-in" datatypes alongside
custom RDFS datatypes (such as UCO's vocabulary namespace members).

The anonymous node shape design pattern is documented in PR 564, commit
77cd47d.

This patch is known to fail CI at the moment because it catches the
error described in 562, but does not resolve that situation.  This patch
is provided now to prevent this implementation error pattern from
arising in the future.

References:
* https://github.com/ucoProject/UCO/issues/562
* https://github.com/ucoProject/UCO/pull/564/commits/77cd47d3e13bce2a6278b8416152b08f5a5d0c32

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>