Revise sequence-application shapes to target predicates' subjects
This causes `ValidationResult`s to yield sufficient information to find
the whole triple that should have used a sequence.  With the prior
spelling, only the object of the triple would have been directly
identified; the predicate could be manually deduced from the shape name;
and the subject was completely absent from the validation result.

The noted comment on PR 572 illustrates the difference in results when
testing CASE-Corpora.  The issue is replicated as new PASS and XFAIL
test members.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ucoProject/UCO/issues/571
* https://github.com/ucoProject/UCO/pull/572#issuecomment-1841238572

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>